+++ src/components/banner/GovBanner/GovBanner.tsx
@@ -10,9 +10,9 @@ import { MediaBlockBody } from '../../mediablock/MediaBlockBody/MediaBlockBody'
 import { Icon } from '../../Icon/Icons'
 
 // assets
-import flagImg from '@uswds/uswds/src/img/us_flag_small.png'
-import dotGovIcon from '@uswds/uswds/src/img/icon-dot-gov.svg'
-import httpsIcon from '@uswds/uswds/src/img/icon-https.svg'
+import flagImg from '@uswds/uswds/img/us_flag_small.png'
+import dotGovIcon from '@uswds/uswds/img/icon-dot-gov.svg'
+import httpsIcon from '@uswds/uswds/img/icon-https.svg'
 
 export type Language = 'english' | 'spanish'