+++ src/components/header/Header/Header.stories.tsx
@@ -84,8 +84,7 @@ export const BasicHeader = (): React.ReactElement => {
 
   return (
     <>
-      <div className={`usa-overlay ${expanded ? 'is-visible' : ''}`}></div>
-      <Header basic={true}>
+      <Header basic={true} showMobileOverlay={expanded}>
         <div className="usa-nav-container">
           <div className="usa-navbar">
             <Title>Project Title</Title>
@@ -191,8 +190,10 @@ export const BasicHeaderWithMegaMenu = (): React.ReactElement => {
 
   return (
     <>
-      <div className={`usa-overlay ${expanded ? 'is-visible' : ''}`}></div>
-      <Header basic={true} basicWithMegaMenu={true}>
+      <Header
+        basic={true}
+        basicWithMegaMenu={true}
+        showMobileOverlay={expanded}>
         <div className="usa-nav-container">
           <div className="usa-navbar">
             <Title>Project Title</Title>
@@ -210,7 +211,7 @@ export const BasicHeaderWithMegaMenu = (): React.ReactElement => {
   )
 }
 
-export const extendedHeader = (): React.ReactElement => {
+export const ExtendedHeader = (): React.ReactElement => {
   const [expanded, setExpanded] = useState(false)
   const onClick = (): void => setExpanded((prvExpanded) => !prvExpanded)
   const [isOpen, setIsOpen] = useState([false])
@@ -251,8 +252,7 @@ export const extendedHeader = (): React.ReactElement => {
   ]
   return (
     <>
-      <div className={`usa-overlay ${expanded ? 'is-visible' : ''}`}></div>
-      <Header extended={true}>
+      <Header extended={true} showMobileOverlay={expanded}>
         <div className="usa-navbar">
           <Title>Project Title</Title>
           <NavMenuButton onClick={onClick} label="Menu" />
@@ -269,7 +269,7 @@ export const extendedHeader = (): React.ReactElement => {
   )
 }
 
-export const extendedHeaderWithMegaMenu = (): React.ReactElement => {
+export const ExtendedHeaderWithMegaMenu = (): React.ReactElement => {
   const [expanded, setExpanded] = useState(false)
   const onClick = (): void => setExpanded((prvExpanded) => !prvExpanded)
 
@@ -330,8 +330,7 @@ export const extendedHeaderWithMegaMenu = (): React.ReactElement => {
   ]
   return (
     <>
-      <div className={`usa-overlay ${expanded ? 'is-visible' : ''}`}></div>
-      <Header extended={true}>
+      <Header extended={true} showMobileOverlay={expanded}>
         <div className="usa-navbar">
           <Title>Project Title</Title>
           <NavMenuButton onClick={onClick} label="Menu" />