+++ src/stories/templates/createaccount.stories.tsx
@@ -30,12 +30,12 @@ import {
   IdentifierGov,
 } from '../../index'
 
-import logoImg from '@uswds/uswds/src/img/logo-img.png'
-import circleSvg from '@uswds/uswds/src/img/circle-gray-20.svg'
+import logoImg from '@uswds/uswds/img/logo-img.png'
+import circleSvg from '@uswds/uswds/img/circle-gray-20.svg'
 import { MediaBlockBody } from '../../components/mediablock/MediaBlockBody/MediaBlockBody'
 
 export default {
-  title: 'Page Templates/Authentication Pages',
+  title: 'Page Templates/Create Account',
   parameters: {
     options: {
       showPanel: false,
@@ -328,7 +328,8 @@ export const CreateAccount = (): React.ReactElement => {
             </IdentifierLogo>
           </IdentifierLogos>
           <IdentifierIdentity domain="domain.gov">
-            An official website of the <Link href="#">{`<Parent agency>`}</Link>
+            <span aria-hidden="true">An</span> official website of the{' '}
+            <Link href="#">{`<Parent agency>`}</Link>
           </IdentifierIdentity>
         </IdentifierMasthead>
         <IdentifierLinks navProps={{ 'aria-label': 'Important links' }}>