+++ src/stories/templates/notfound.stories.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React, { useState } from 'react'
-import logoImg from '@uswds/uswds/src/img/logo-img.png'
-import circleSvg from '@uswds/uswds/src/img/circle-gray-20.svg'
+import logoImg from '@uswds/uswds/img/logo-img.png'
+import circleSvg from '@uswds/uswds/img/circle-gray-20.svg'
 
 import {
   GovBanner,
@@ -35,7 +35,7 @@ import {
 } from '../../index'
 
 export default {
-  title: 'Page Templates/Not Found Page',
+  title: 'Page Templates/Not Found',
   parameters: {
     options: {
       showPanel: false,
@@ -187,8 +187,7 @@ export const NotFoundPage = (): React.ReactElement => {
         Skip to main content
       </a>
       <GovBanner />
-      <div className={`usa-overlay ${mobileNavOpen ? 'is-visible' : ''}`}></div>
-      <Header basic>
+      <Header basic showMobileOverlay={mobileNavOpen}>
         <div className="usa-nav-container">
           <div className="usa-navbar">
             <Title id="basic-logo">
@@ -287,7 +286,7 @@ export const NotFoundPage = (): React.ReactElement => {
             </IdentifierLogo>
           </IdentifierLogos>
           <IdentifierIdentity domain={'<domain.gov>'}>
-            An official website of the{' '}
+            <span aria-hidden="true">An</span> official website of the{' '}
             <Link href="javascript:void(0);">{`<Parent agency>`}</Link>
           </IdentifierIdentity>
         </IdentifierMasthead>