+++ src/uswds.ts
@@ -1,1 +1,0 @@
-import '@uswds/uswds/css/uswds.css'